refactoring:add image to chat
