+++ telegram_framework/messages/core/__init__.py
@@ -1,2 +1,3 @@
 from .core_message_base import CoreMessageBase
 from .core_message_default import CoreMessageDefault
+from .chat_message import create_chat_message, is_chat_message
