+++ telegram_framework/messages/reply/reply.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import Any
 from telegram_framework.messages.errors import MessageNotInChatError
 from telegram_framework.messages.text_message import MessageDefault
-from telegram_framework.messages.chat_message import is_chat_message
+from telegram_framework.messages.core import is_chat_message
 from .reply_base import ReplyBase
